Make the README and NOTES more understandable and organized; Promote the role of the the pg pipeline, now that it is used to download multiple tables
