+++ Procfile
@@ -2,7 +2,7 @@ friend_collector: python -m app.workers.friend_batch_collector
 bq_grapher: python -m app.workers.bq_grapher
 bq_list_grapher: python -m app.workers.bq_list_grapher
 pg_migrate: python -m app.models
-pg_pipeline: python -m app.workers.pg_pipeline
+pg_pipeline_user_friends: python -m app.workers.pg_pipeline_user_friends
 pg_pipeline_user_details: python -m app.workers.pg_pipeline_user_details
 pg_pipeline_retweeter_details: python -m app.workers.pg_pipeline_retweeter_details
 psycopg_list_grapher: python -m app.workers.psycopg_list_grapher
