+++ app/workers/pg_pipeline_retweeter_details.py
@@ -1,6 +1,6 @@
 
 
-from app.workers.pg_pipeline import Pipeline
+from app.pg_pipeline import Pipeline
 
 if __name__ == "__main__":
 
@@ -9,3 +9,5 @@ if __name__ == "__main__":
     pipeline.download_retweeter_details() # takes about 16 minutes for 2.7M users in batches of 2500
 
     pipeline.report()
+
+    pipeline.sleep()
