+++ app/workers/pg_pipeline_user_friends.py
@@ -6,7 +6,7 @@ if __name__ == "__main__":
 
     pipeline = Pipeline()
 
-    pipeline.download_user_details() # takes about 50 minutes for 3.6M users in batches of 2500
+    pipeline.download_user_friends()
 
     pipeline.report()
